[Mosaic:TPU] Efficient relayout with internal scratch
We should already support all different retilings with any dtype in this cl at the moment. The efficient relayout with scratch brings significant improvements on current retiling in <= TPUv4 and retiling with (packing, 128) in TPUv5. All missing retiling supports are added in this cl, including increase sublane retiling and packed type retiling.

PiperOrigin-RevId: 651670663